Remove Services/Dashboard package; WMRuntime DashboardMonitor and DashboardInterface
Apply Todors suggestions
