--- src/python/WMCore/Services/Dashboard/__init__.py
@@ -1,8 +1,0 @@
-#!/usr/bin/env python
-"""
-_WMCore/Services_
-
-Services for Workload Management Packages
-
-"""
-__all__ = []
